refactor: phoneNumber 입력 컴포넌트 payment폴더로 이동
